pyproject.toml
[project]
name = "not-sure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.78.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.10",
    "scipy>=1.15.3",
    "tiktoken>=0.9.0",
]
